+++ goCat/Features/Home/ViewModels/HomeViewModel.swift
@@ -4,8 +4,6 @@ import Observation
 @MainActor
 @Observable
 final class HomeViewModel {
-    var selectedCat = CatOption.starter
-    var selectedChair = ChairOption.starter
     var selectedScene = SceneOption.study
     var selectedSound = SoundOption.rain
     var sessionDuration: TimeInterval = 25 * 60